Move SupportedNetworks to separate file for better enum support
