+++ src/renderer/services/eth2/supportedNetworks.ts
@@ -1,0 +1,3 @@
+export enum SupportedNetworks {
+    PRYSM = "Prysm",
+}
