+++ test/e2e/redux/reducers/register.reducer.spec.ts
@@ -10,7 +10,7 @@ import {
 import {IRegisterState, registerReducer as reducer} from "../../../../src/renderer/reducers/register";
 import {RegisterActionTypes} from "../../../../src/renderer/constants/action-types";
 import {Action} from "redux";
-import {SupportedNetworks} from "../../../../src/renderer/services/docker/chain";
+import {SupportedNetworks} from "../../../../src/renderer/services/eth2/supportedNetworks";
 const privateKeyStr = "0xd68ffdb8b9729cb02c5be506e9a2fad086746b4bdc2f50fb74d10ac8419c5259";
 const publicKeyStr =
     "0x92fffcc44e690220c190be41378baf6152560eb13fa73bdf8b45120b56096acc4b4e87a0e0b97f83e48f0ff4990daa18";
