+++ test/unit/eth2/client/prysm/validator.spec.ts
@@ -1,5 +1,5 @@
 import {networks} from "../../../../../src/renderer/services/eth2/networks";
-import {SupportedNetworks} from "../../../../../src/renderer/services/docker/chain";
+import {SupportedNetworks} from "../../../../../src/renderer/services/eth2/supportedNetworks";
 import axios from "axios";
 import MockAxiosAdapter from "@nodefactory/axios-mock-adapter";
 import {
